+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
-## [Unreleased](https://github.com/USGS-WiM/StreamStats/tree/dev)
+## [Unreleased](https://github.com/USGS-WiM/StreamStats/tree/SC-Runoff)
   
   ### Added
 
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     -  Query NHD streams in South Carolina and add new section to report
 
   ### Changed
+
+    - Remove Hydrologic Features sidebar button and modal; show all intersecting NHD streams to report
 
   ### Fixed